Refactor: logger, utils를 libs의 utils로 이동
