+++ apps/api/src/app.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get } from '@nestjs/common';
 import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Public } from '@api/auth/auth.decorator';
-import { logger } from '@api/config/logger';
+import { logger } from '@app/utils/logger';
 
 @ApiTags('Hello')
 @Controller()
