+++ apps/api/src/article/repositories/article.repository.ts
@@ -3,7 +3,7 @@ import { Article } from '@api/article/entities/article.entity';
 import { FindAllArticleRequestDto } from '@api/article/dto/request/find-all-article-request.dto';
 import { NotFoundException } from '@nestjs/common';
 import { PaginationRequestDto } from '@api/pagination/dto/pagination-request.dto';
-import { getPaginationSkip } from '@api/utils';
+import { getPaginationSkip } from '@app/utils/utils';
 
 @EntityRepository(Article)
 export class ArticleRepository extends Repository<Article> {
