+++ apps/api/src/comment/repositories/comment.repository.ts
@@ -2,7 +2,7 @@ import { NotFoundException } from '@nestjs/common';
 import { EntityRepository, Repository } from 'typeorm';
 import { Comment } from '@api/comment/entities/comment.entity';
 import { PaginationRequestDto } from '@api/pagination/dto/pagination-request.dto';
-import { getPaginationSkip } from '@api/utils';
+import { getPaginationSkip } from '@app/utils/utils';
 
 @EntityRepository(Comment)
 export class CommentRepository extends Repository<Comment> {
