+++ apps/api/src/intra-auth/intra-auth.controller.ts
@@ -22,7 +22,7 @@ import { GetUser, OnlyNovice, Public } from '@api/auth/auth.decorator';
 import { User } from '@api/user/entities/user.entity';
 import { SigninIntraAuthRequestDto } from './dto/signin-intra-auth-request.dto';
 import { AllExceptionsFilter } from '@app/common/filters/all-exception.filter';
-import { logger } from '@api/config/logger';
+import { logger } from '@app/utils/logger';
 
 @ApiTags('Intra Auth')
 @Controller('intra-auth')
