+++ apps/api/test/e2e/jest-e2e.json
@@ -9,6 +9,7 @@
   "moduleNameMapper": {
     "^@api/(.*)$": "<rootDir>/../../src/$1",
     "^@test/(.*)$": "<rootDir>/../../test/$1",
-    "^@app/common/(.*)$": "<rootDir>/../../../../libs/common/src/$1"
+    "^@app/common/(.*)$": "<rootDir>/../../../../libs/common/src/$1",
+    "^@app/utils/(.*)$": "<rootDir>/../../../../libs/utils/src/$1"
   }
 }
