+++ libs/common/src/cache/cache.service.ts
@@ -1,7 +1,7 @@
 import { CACHE_MANAGER, Inject, Injectable } from '@nestjs/common';
 import { Cache, CachingConfig } from 'cache-manager';
 
-import { TIME2LIVE } from '@api/utils';
+import { TIME2LIVE } from '@app/utils/utils';
 import { IntraAuthMailDto } from '@app/common/cache/dto/intra-auth.dto';
 
 @Injectable()
