+++ tsconfig.json
@@ -11,15 +11,10 @@
     "outDir": "./dist",
     "baseUrl": "./",
     "paths": {
-      "@api/*": [
-        "./apps/api/src/*"
-      ],
-      "@test/*": [
-        "./apps/api/test/*"
-      ],
-      "@app/common/*": [
-        "./libs/common/src/*"
-      ]
+      "@api/*": ["./apps/api/src/*"],
+      "@test/*": ["./apps/api/test/*"],
+      "@app/common/*": ["./libs/common/src/*"],
+      "@app/utils/*": ["./libs/utils/src/*"]
     },
     "incremental": true,
     "skipLibCheck": true,
